+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `ovni_version_get()` function.
-- Add the `ovniver` program to report the `libovni.so` version.
+- Add the `ovniver` program to report the libovni version and commit.
 
 ## [1.2.2] - 2022-07-26